More warnings for Win64 removed.
git-svn-id: http://blosc.pytables.org/svn/blosc/trunk@258 5b11b6af-1e7b-4319-8082-0aacbd7ffead

+++ blosc/shuffle.c
@@ -236,7 +236,7 @@ shuffle16(uint8_t* dest, uint8_t* src, size_t size)
 /* Shuffle a block.  This can never fail. */
 void shuffle(size_t bytesoftype, size_t blocksize,
              uint8_t* _src, uint8_t* _dest) {
-  int unaligned_dest = (uintptr_t)_dest % 16;
+  int unaligned_dest = (int)((uintptr_t)_dest % 16);
   int power_of_two = (blocksize & (blocksize - 1)) == 0;
   int too_small = (blocksize < 256);
 
@@ -449,8 +449,8 @@ unshuffle16(uint8_t* dest, uint8_t* orig, size_t size)
 /* Unshuffle a block.  This can never fail. */
 void unshuffle(size_t bytesoftype, size_t blocksize,
                uint8_t* _src, uint8_t* _dest) {
-  int unaligned_src = (uintptr_t)_src % 16;
-  int unaligned_dest = (uintptr_t)_dest % 16;
+  int unaligned_src = (int)((uintptr_t)_src % 16);
+  int unaligned_dest = (int)((uintptr_t)_dest % 16);
   int power_of_two = (blocksize & (blocksize - 1)) == 0;
   int too_small = (blocksize < 256);
